Change coordinates of triangle to equilateral triangle
Existing triangle is based on a unit square, with two sides longer than the third. New coordinates are based on the halfway point between a triangle inscribed within a unit circle and a triangle circumscribing a unit circle. Radius = 20 will get y = 30 at the top, but visually it will look of similar size to a circle of radius = 20.